add support for excluding modules for object prefix and allow no input on mutations
