+++ CHANGELOG.md
@@ -4,8 +4,8 @@
 
 - Added checks
   - `AbsintheMutationInput`
-  - `AbsintheMutationUniqueObject`
-  - `AbsintheObjectName`
+  - `AbsintheMutationPayload`
+  - `AbsintheObjectPrefix`
 - General documentation and tooling
 
 ## v0.1.0 (YYYY-MM-DD)
